separated classes and specs to individual files, updated tests to use , simplified methods in each class from PR feedback
